requirements.txt
flask==3.1.2
gunicorn==25.1.0
line-bot-sdk==3.21.0
supabase==2.28.0
requests==2.32.5
beautifulsoup4==4.12.3
python-dotenv==1.2.1
openai>=1.0.0